changed math function tan  calls to tanf

+++ src/camera.c
@@ -46,7 +46,7 @@ bool init_camera_for_frame(t_minirt *minirt, t_camera *cam)
 	const t_v3 w = unit_vector(V3_SUB(cam->lookfrom, cam->lookat));
 	const t_v3 u = unit_vector(cross(cam->vup, w));
 	const t_v3 v = cross(w, u);
-	const float h = (float)tan((cam->fov * M_PI / 180) / 2);
+	const float h = tanf((cam->fov * M_PI / 180) / 2);
 
 	cam->pixel_sample_scale = 1.0f / cam->samples_per_pixel;
 	cam->aspect_ratio = (float)minirt->image->width / minirt->image->height;
@@ -69,7 +69,7 @@ bool init_camera_for_frame(t_minirt *minirt, t_camera *cam)
 
 	cam->viewport_upper_left = viewport_top_left(cam, w);
 
- 	const float defocus_radius = cam->focus_dist * tan((cam->defocus_angle / 2) * M_PI / 180);
+ 	const float defocus_radius = cam->focus_dist * tanf((cam->defocus_angle / 2) * M_PI / 180);
   	cam->defocus_disk_u = v3_mul_f32(u, defocus_radius);
     cam->defocus_disk_v = v3_mul_f32(v, defocus_radius);
